renaming EUDET TLU
