+++ user/eudet/hardware/CMakeLists.txt
@@ -6,19 +6,20 @@ if(USER_BUILD_EUDET_TLU)
   include_directories(${LIBUSB_INCLUDE_DIRS})
   add_definitions(${TLUFIRMWARE_DEFINITIONS})
   list(APPEND USER_HARDWARE_SRC
-    src/TLUController.cc
-    src/TLUAddresses1.cc
-    src/TLUAddresses2.cc)
+    src/EudetTluController.cc
+    src/EudetTluAddresses1.cc
+    src/EudetTluAddresses2.cc)
   list(APPEND USER_HARDWARE_DEP_LIB ${LIBUSB_LIBRARIES} ${ZESTSC1_LIBRARIES})
 endif()
 
 if(USER_BUILD_AIDA_TLU)
   include_directories(${CACTUS_INCLUDE_DIR})
-  list(APPEND USER_HARDWARE_SRC src/AidaTluController.cc)
-  list(APPEND USER_HARDWARE_SRC src/AidaTluHardware.cc)
-  list(APPEND USER_HARDWARE_SRC src/AidaTluPowerModule.cc)
-  list(APPEND USER_HARDWARE_SRC src/AidaTluI2c.cc)
-  list(APPEND USER_HARDWARE_SRC src/AidaTluDisplay.cc)
+  list(APPEND USER_HARDWARE_SRC 
+    src/AidaTluController.cc
+    src/AidaTluHardware.cc
+    src/AidaTluPowerModule.cc
+    src/AidaTluI2c.cc
+    src/AidaTluDisplay.cc)
   list(APPEND USER_HARDWARE_DEP_LIB ${CACTUS_LIBRARIES})
 endif()
