+++ user/eudet/hardware/src/EudetTluAddresses2.cc
@@ -1,5 +1,5 @@
-#include "TLUAddresses.hh"
-#include "TLU_address_map_v0-2.h"
+#include "EudetTluAddresses.hh"
+#include "EudetTlu_address_map_v0-2.h"
 
 #define BEAM_TRIGGER_FSM_RESET_BIT 4
 #define DMA_CONTROLLER_RESET_BIT 5
