+++ user/eudet/misc/starting_scripts/02_eudet_tlu_telescope
@@ -15,7 +15,7 @@ xterm -T "Data Collector NI/Mimosa" -e 'euCliCollector -n DirectSaveDataCollecto
 sleep 1
 
 # Start TLU Producer
-xterm -T "EUDET TLU Producer" -e 'euCliProducer -n TluProducer -t eudet_tlu -r tcp://192.168.22.1:44000' & 
+xterm -T "EUDET TLU Producer" -e 'euCliProducer -n EudetTluProducer -t eudet_tlu -r tcp://192.168.22.1:44000' & 
 
 # Start NI Producer locally connect to LV via TCP/IP
 xterm -T "NI/Mimosa Producer" -e 'euCliProducer -n NiProducer -t ni_mimosa -r tcp://192.168.22.1:44000' &
